Update check_psspy_already_in_path().
This changes it from trying to import psspy to see if the PSSBIN is already on
the system path. This speeds up execution and makes the program more flexible.
We are now able to check to see if there is a mismatch between the paths in
sys.path and os.environ['PATH'].